Add helper function for tests to create provider and delegator MSAs
